+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.0 (unreleased)
 
+- Dropped support for marshal serialization
 - Dropped support for Ruby < 3.1 and Rails < 7
 
 ## 0.4.2 (2024-06-24)